pyproject.toml
[project]
name = "weather-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[openai,sarvam,silero]>=0.12.0",
    "livekit-plugins-openai>=1.5.0",
    "livekit-plugins-sarvam>=1.5.0",
    "python-dotenv>=1.2.2",
    "sarvamai>=0.1.27",
]
